Move `applyServerChanges` from component.js to utils.js
